feat: Add Tailwind CSS configuration and plugins
This commit adds the necessary configuration files for integrating Tailwind CSS into the project. The `postcss.config.js` file is created with the required plugins for Tailwind CSS and Autoprefixer. Additionally, the `tailwind.config.js` file is added to define the content and theme options for Tailwind CSS.

The following files are modified:
- `postcss.config.js`
- `tailwind.config.js`

The following files are added:
- `src/components/SimplePage.jsx`
- `src/hooks/axios.js`
- `src/components/Component.jsx`
- `src/hooks/user/auth.js`
- `src/routes.jsx`
- `src/components/Alert.jsx`
- `src/components/Typography.jsx`
- `src/pages/Logout.jsx`
- `src/components/Navbar.jsx`

The following dependencies are added:
- `@tabler/icons-react@3.5.0`
- `axios@1.7.2`
- `dayjs@1.11.11`
- `jwt-decode@4.0.0`
- `react-hook-form@7.51.5`
- `react-router-dom@6.23.1`

The following dev dependencies are added:
- `autoprefixer@10.4.19`
- `postcss@8.4.38`
- `tailwindcss@3.4.3`

Please review and merge this commit.